Improve YouTube recent video check
Previous method was failing to update videos. Instead of a cutoff this mehtod just removes any previously gotten videos from the possibility list.
This does mean that if Reddit posting fails, and does not crash the script, the old position will be saved and videos will get missed. Fixing this should be explored in the future, especially if we plan on catching more errors instead of letting the script crash.